+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to the **ValidationExtensions** will be documented in this file. The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ * Added `RequiredDynamicAttribute` which can accept method to validate complicated requirement(s) and more readable code.
+
+## [Unreleased]
 
 ### Added
  * Removed all `*IfNotAttribute` in favor of newly introducded `IsInverse` parameter. For example: `[RequiredIfNot(nameof(OtherProperty), false)]` --> `[RequiredIf(nameof(OtherProperty), false, isInverse: true)]`